# Faster metal compiled kernels + some fixes