--- a/mlx/backend/metal/kernels/unary.metal
+++ b/mlx/backend/metal/kernels/unary.metal
@@ -8,8 +8,7 @@
 #define instantiate_unary_all(op, tname, type)                  \
   instantiate_kernel("v_" #op #tname, unary_v, type, op)        \
   instantiate_kernel("v2_" #op #tname, unary_v2, type, op)      \
-  instantiate_kernel("gn4_" #op #tname, unary_g, type, op, 4)   \
-  instantiate_kernel("g_" #op #tname, unary_g, type, op)
+  instantiate_kernel("gn4_" #op #tname, unary_g, type, op, 4)
 
 #define instantiate_unary_float(op)               \
   instantiate_unary_all(op, float16, half)        \