- More work per thread for ndim > 3 compiled kernels to amortize cost of indexing
- Fix compile for non-contiguous large array sizes (> uint32_max). Add test to long running CI
- Speed up indexing computations for dynamic dim case (ndim > 8)
- Always use work per thread > 1 for non contiguous elementwise ops (unary, binary, copy, etc). Not sure why I had the dim=1 specialization in there. It's always faster when the last dimension > 1 and collapsing should remove the case when the last dimension is 1. (Added benefit it reduces binary size by 1MB)
- Fix bug in compile for ndim > 8 (dynamic dims) + add test

### Benchmarks

On M1 Max

Bench | Pre (ms) | Post (ms)
---- | ---- | -----
Strided compile dim=6 | 22.961 | 9.387
Strided compile dim=9 | 476.715 | 213.611
Binary last dim = 2 | 3.324 | 2.007
Binary last dim = 4 | 6.387 | 2.683

Code for benchmarks:

```python
D = 24
a = mx.random.uniform(shape=(D,)*6)
a = a.transpose(1, 0, 3, 2, 5, 4)
b = mx.random.uniform(shape=(D,)*6)
mx.eval(a, b)


def fn(a, b):
    return mx.negative(a + b)

timeit(mx.compile(fn), a, b)

D = 8
a = mx.random.uniform(shape=(D,)*9)
a = a.T
b = mx.random.uniform(shape=(D,)*9)
mx.eval(a, b)

def fn(a, b):
    return mx.negative(a + b)

timeit(mx.compile(fn), a, b)

D = 28
a = mx.random.uniform(shape=(D,)*4 + (2, D))
a = a.transpose(1, 0, 3, 2, 5, 4)
b = mx.random.uniform(shape=(D,)*5 + (2,))
mx.eval(a, b)

def fn(a, b):
    return a + b

timeit(fn, a, b)

D = 28
a = mx.random.uniform(shape=(D,)*4 + (4, D))
a = a.transpose(1, 0, 3, 2, 5, 4)
b = mx.random.uniform(shape=(D,)*5 + (4,))
mx.eval(a, b)

def fn(a, b):
    return a + b

timeit(fn, a, b)
```